fix(report_sonar): resolve remaining comments
