+++ tools/devsecops_engine_tools/engine_utilities/sonarqube/src/infrastructure/helpers/utils.py
